+++ client/src/shared/types/common.ts
@@ -1,0 +1,18 @@
+export type Nullable<T> = T | null;
+
+export interface PaginationDTO {
+  page: number;
+  pageSize: number;
+}
+
+export interface Pagination extends PaginationDTO {
+  pageCount: number;
+  total: number;
+}
+
+export interface APIResponse<D> {
+  data: D;
+  meta: {
+    pagination: Pagination;
+  };
+}